Preparing to move to XeLaTeX
XeTeX allows full UTF-8 input, while LaTeX is based on 8-bit
characters. UTF-8 is needed for special symbols (such as those for the
planets), and is also useful for all the languages used.
Kepler Fonts does not work with XeTeX, but it does not need to. We can
use
Features from OpenType/TrueType/similar fonts, such as Hoefler Text.
These fonts can do everything that the Kepler Fonts package does.
Added rest of the text from RTF Pars 1 (which has planet symbols and a
table in it).