Tables Lib.1: Cleaned out old test frames
Also: Filled in standard headers and caption/label

Longtable p043: no label yet.
Fixed all 3 references in chapter file.

NB: The editor needs to show a line number after the last line, or
there is no newline on the last line.
Without it, a % on the last line may comment out the next line in the
main file.