Developing command for Greek numbers
Issue #171

\ruleover{string}
\ruleover[scale]{string}

Put a bar over any string, like \overline, but outside math mode.
The string is shrunk to 0.75 of its original size and a rule
is drawn over the string at the original height of the string.

An alternative scale factor can be given as an optional parameter.
Scales outside the range 0.5-1.0 do not look nice.
This option is intended to allow the user to make the ga[ between
the string and the rule look aesthetically pleasing.
Scaling has no effect on the vertical position of the rule.
The thickness of the rule follows the scale, to match the "fatness"
of the scaled font.

Uses the graphicx package for the \scalebox command.

The \ruleover{} command works outside math mode and tolerates any
kind of font modification such as Greek characters or
small caps.

Internal variables use the prefix "ROver@".

A demonstration of how to implement a \gnum{} command is included.