Command for em-dash: \emd
The em-dash often occurs in quotes.
We could use a litteral unicode em-dash (—), but that is hard to
distinguish from other dashes in the source code.
We could also use the LaTeX command "\textemdash", but that is a bit
verbose.
The TeX "---" code does not work in all fonts (particularly our Greek
font).

All occurences of the unicode em-dash, "---" and "--" replaced with
this command.