Hoefler Text Monospaced and Proportional are reversed
Weirdly Numbers=Monospaced and Numbers=Proportional appear to be
reversed for Hoefler Text.

When Numbers=Proportional is set, the numbers are monospaced (useful
for tables) with each digit taking up the same width.

When Numbers=Monospaced is set, the numbers are proportional, with
different digits taking up a different width.