Tables p139: Tweaked width
An Overfull \hbox was reported. Solved by adjusting various widths.
Replaced spacing with 'p' column by '@{\hspace{...}}'
Because minipage adds some margin arounds itself, the sum of the sizes
must be less than 1\textwidh. As we had 0.68+0.04+0.28=1.00 the right
table would bleed over into the right margin, leading to the Overfull
\hbox warning.
The sum is now 0.65+0.05+0.27=0.97, which gives appears to be good
enough.
Secondary: Width of the right-hand table was determined by the width of
the header text, causing the distance between the rightmost column and
the rest to be stretched.