Tweaked titlespacing to match page length
Note: titlespacing without 'plus' and 'minus' is in underful-vbox-02 branch.
Increased the default size until the bottom line touches the bottom margin (using \raggedbottom), while the last line does not jump over to the next page.
For the two-line title the space between the last line and the bottom margin has become a lot smaller too.