Raised the hight of the 'rule' to the next line hight
New: 4.914 ex. Old: 2.227 ex. Difference: 2.687 ex. This is thus approximately equal to the \backlineskip value (the distance from one line's baseline to the next) which is printed in the output as being 14.5 points.
Default flexibility of the titlespacing is 1.5ex (maximum stretched size 7.0ex, minimal shrunk size 5.5ex). This must be raised to at least  2.227ex  to let it handle all situations.